requirements.txt
accelerate==1.5.2
av==14.2.0
boto3==1.37.27
botocore==1.37.27
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
olefile==0.47
packaging==24.2
pillow==11.0.0
psutil==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.4
safetensors==0.5.3
setuptools==75.8.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
